--wip-- src: lib: lore: create lore library
This gets the mailing lists available at the lore archive, and will be
used to create the lore interface in the future.

To do this it relies on a sed script that parses the HTML web page
the sed scripts used are as follows:

    s/^href="(.*)\/?">\1<\/a>$/\1/p

  This one identifies the line that contains the title of the mailing
list, captures said title and returns it.

    s/^  (.*)$/\1/p

  This second identifies the line that contains the description of the
list and returns it.

When these commands are executed in sequence (separated by a colon),
magic happens and the title and description are returned in order. This
result is then used to store these values in a global associative array.

Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>